refactor(coach): isolate trace and session funnel helpers
