+++ src/app/api/coach/route.ts
@@ -15,7 +15,6 @@ import { getCoachRuntime } from "@/lib/coach/server/runtime";
 import {
   buildEndOfTurnSuggestions,
   runPlannerTurn,
-  type PlannerRunResult,
 } from "@/lib/coach/server/planner";
 import {
   extractMemoryOperations,
@@ -23,9 +22,10 @@ import {
   summarizeObservation,
   type MemoryTranscriptMessage,
 } from "@/lib/coach/server/memory-pipeline";
+import { buildCoachTraceData } from "@/lib/coach/server/trace";
 import { buildRuntimeUnavailableResponse } from "@/lib/coach/server/blocks";
 import { sanitizeError } from "@/lib/coach/sanitize-error";
-import type { CoachTraceData, CoachUIMessage } from "@/lib/coach/ui-message";
+import type { CoachUIMessage } from "@/lib/coach/ui-message";
 import {
   generateText,
   convertToModelMessages,
@@ -55,57 +55,6 @@ type CoachContextState = {
   summary: string | null;
   messages: StoredCoachMessage[];
 };
-
-function countUserTurns(history: ModelMessage[]): number {
-  return Math.max(
-    0,
-    history.filter((message) => message.role === "user").length - 1
-  );
-}
-
-function extractFirstLoggedExercise(
-  plannerResult: PlannerRunResult
-): string | null {
-  for (const result of plannerResult.toolResults) {
-    if (result.toolName !== "log_sets") continue;
-
-    const directExercise = result.outputForModel.exercise_name;
-    if (typeof directExercise === "string" && directExercise.trim()) {
-      return directExercise;
-    }
-
-    const nestedResults = result.outputForModel.results;
-    if (!Array.isArray(nestedResults)) continue;
-
-    for (const nestedResult of nestedResults) {
-      if (!nestedResult || typeof nestedResult !== "object") continue;
-      const exerciseName = (nestedResult as Record<string, unknown>)
-        .exercise_name;
-      if (typeof exerciseName === "string" && exerciseName.trim()) {
-        return exerciseName;
-      }
-    }
-  }
-
-  return null;
-}
-
-function buildCoachTraceData({
-  sessionId,
-  history,
-  plannerResult,
-}: {
-  sessionId?: string;
-  history: ModelMessage[];
-  plannerResult: Awaited<ReturnType<typeof runPlannerTurn>>;
-}): CoachTraceData {
-  return {
-    session_id: sessionId ?? null,
-    tool_calls_count: plannerResult.toolsUsed.length,
-    turn_index: countUserTurns(history),
-    first_logged_exercise: extractFirstLoggedExercise(plannerResult),
-  };
-}
 
 function deserializeStoredMessage(content: string): ModelMessage | null {
   try {
